backend/pyproject.toml
[tool.poetry]
name = "lola-backend"
version = "0.1.0"
description = "Backend of LOLA, a classification system for flamenco styles."
authors = ["Ale Sanchez <alesanchezmedina@gmail.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
Flask = "^2.0.1"
python-dotenv = "^0.19.0"
requests = "^2.26.0"
gunicorn = "^20.1.0"
tensorflow-hub = "^0.12.0"
numpy = "^1.19.5"
joblib = "^1.1.0"
pydub = "^0.25.1"
tensorflow = "^2.6.0"
scikit-learn = "^1.0.1"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"
pylint = "^2.9.6"
flake8 = "^3.9.2"
bandit = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
